# docs/swagger.yaml
definitions:
  types.Code:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeRunning
    - CodeFailed
    - CodeNoData
    - CodePending
    - CodePaused
  types.SBase-any:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data: {}
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-array_types_STaskStepRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        items:
          $ref: '#/definitions/types.STaskStepRes'
        type: array
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
...
  types.SBase-types_SPool:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        $ref: '#/definitions/types.SPool'
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-types_STaskCreateRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        $ref: '#/definitions/types.STaskCreateRes'
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-types_STaskDetailRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        $ref: '#/definitions/types.STaskDetailRes'
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-types_STaskListDetailRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        $ref: '#/definitions/types.STaskListDetailRes'
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-types_STaskStepLogListRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        items:
          $ref: '#/definitions/types.STaskStepLogRes'
        type: array
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
  types.SBase-types_STaskStepRes:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/types.Code'
        example: 255
      data:
        $ref: '#/definitions/types.STaskStepRes'
      msg:
        example: message
        type: string
      timestamp:
        type: integer
    type: object
...
  types.SPageRes:
    properties:
      current:
        description: 当前页
        type: integer
      size:
        description: 页大小
        type: integer
      total:
        description: 总页数
        type: integer
    type: object
  types.SPool:
    properties:
      size:
        example: 30
        type: integer
    required:
    - size
    type: object
  types.STaskCreateRes:
    properties:
      count:
        description: 步骤数量
        type: integer
      name:
        description: 任务名称
        type: string
    type: object
  types.STaskDetailRes:
    properties:
      count:
        description: 步骤数量
        type: integer
      description:
        description: 任务描述
        type: string
      disable:
        description: 是否禁用
        type: boolean
      env:
        additionalProperties:
          type: string
        description: 任务环境变量
        type: object
      msg:
        description: 任务信息
        type: string
      name:
        description: 任务名称
        type: string
      node:
        description: 节点名称
        type: string
      state:
        description: 任务状态
        type: string
      time:
        allOf:
        - $ref: '#/definitions/types.STimeRes'
        description: 时间
      timeout:
        description: 任务超时时间
        type: string
    type: object
  types.STaskListDetailRes:
    properties:
      page:
        allOf:
        - $ref: '#/definitions/types.SPageRes'
        description: 分页
      tasks:
        description: 任务列表
        items:
          $ref: '#/definitions/types.STaskDetailRes'
        type: array
    type: object
  types.STaskReq:
    properties:
      async:
        description: 是否异步执行
        example: false
        type: boolean
      description:
        description: 任务描述
        type: string
      disable:
        description: 是否禁用
        example: false
        type: boolean
      env:
        additionalProperties:
          type: string
        description: 任务环境变量
        example:
          key: value
          key1: value1
        type: object
      name:
        description: 任务名称
        example: task_name
        type: string
      node:
        description: Node 执行节点
        example: ""
        type: string
      step:
        description: 任务步骤
        items:
          $ref: '#/definitions/types.STaskStepReq'
        type: array
      timeout:
        description: 任务超时时间
        example: 24h
        type: string
    type: object
  types.STaskStepLogRes:
    properties:
      content:
        description: 内容
        type: string
      line:
        description: 行号
        type: integer
      timestamp:
        description: 时间戳
        type: integer
    type: object
  types.STaskStepReq:
    properties:
      content:
        description: 步骤内容
        example: sleep 10
        type: string
      depends:
        description: 步骤依赖
        example:
        - ""
        items:
          type: string
        type: array
      description:
        description: 步骤描述
        example: script.ps1
        type: string
      disable:
        description: 是否禁用
        example: false
        type: boolean
      env:
        additionalProperties:
          type: string
        description: 步骤环境变量
        example:
          key: value
          key1: value1
        type: object
      name:
        description: 步骤名称
        example: script.ps1
        type: string
      timeout:
        description: 步骤超时时间
        example: 3m
        type: string
      type:
        description: 步骤超时时间
        example: powershell
        type: string
    type: object
  types.STaskStepRes:
    properties:
      code:
        description: 步骤退出码
        type: integer
      content:
        description: 步骤内容
        type: string
      depends:
        description: 步骤依赖
        items:
          type: string
        type: array
      description:
        description: 步骤描述
        type: string
      disable:
        description: 是否禁用
        type: boolean
      env:
        additionalProperties:
          type: string
        description: 步骤环境变量
        type: object
      msg:
        description: 步骤信息
        type: string
      name:
        description: 步骤名称
        type: string
      state:
        description: 步骤状态
        type: string
      time:
        allOf:
        - $ref: '#/definitions/types.STimeRes'
        description: 时间
      timeout:
        description: 步骤超时时间
        type: string
      type:
        description: 步骤类型
        type: string
    type: object
  types.STimeRes:
    properties:
      end:
        description: 结束时间
        type: string
      start:
        description: 开始时间
        type: string
    type: object
...
info:
  contact:
    name: AutoExecFlow
    url: https://github.com/xmapst/AutoExecFlow/issues
  description: |-
    An `API` for cross-platform custom orchestration of execution steps
    without any third-party dependencies. Based on `DAG`, it implements the scheduling
    function of sequential execution of dependent steps and concurrent execution of
    non-dependent steps. <br /><br /> It provides `API` remote operation mode, batch
    execution of `Shell` , `Powershell` , `Python` and other commands, and easily
    completes common management tasks such as running automated operation and maintenance
    scripts, polling processes, installing or uninstalling software, updating applications,
    and installing patches.
  license:
    name: GPL-3.0
    url: https://github.com/xmapst/AutoExecFlow/blob/main/LICENSE
  title: Auto Exec Flow
  version: "1.0"
paths:
...
  /api/v1/pool:
    get:
      consumes:
      - application/json
      description: 获取工作池信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_SPool'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 详情
      tags:
      - 工作池
    post:
      consumes:
      - application/json
      description: 设置工作池大小
      parameters:
      - description: pool setting
        in: body
        name: setting
        required: true
        schema:
          $ref: '#/definitions/types.SPool'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_SPool'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 设置
      tags:
      - 工作池
...
  /api/v1/task:
    get:
      consumes:
      - application/json
      description: 获取任务列表, 支持WS长连接
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 100
        description: 分页大小
        in: query
        name: size
        type: integer
      - description: 关键字
        in: query
        name: prefix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskListDetailRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 列表
      tags:
      - 任务
    post:
      consumes:
      - application/json
      description: 创建任务
      parameters:
      - description: 任务名称
        in: query
        name: name
        type: string
      - default: false
        description: 是否异步执行
        in: query
        name: async
        type: boolean
      - description: 执行超时时间
        in: query
        name: timeout
        type: string
      - collectionFormat: csv
        description: 任务环境变量
        in: query
        items:
          type: string
        name: env
        type: array
      - description: 步骤内容
        in: body
        name: steps
        required: true
        schema:
          items:
            $ref: '#/definitions/types.STaskStepReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskCreateRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 创建
      tags:
      - 任务
  /api/v1/task/{task}:
    delete:
      consumes:
      - application/json
      description: 删除指定任务
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 删除
      tags:
      - 任务
    get:
      consumes:
      - application/json
      description: 获取指定任务的步骤列表, 支持WS长连接
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-array_types_STaskStepRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 列表
      tags:
      - 步骤
    put:
      consumes:
      - application/json
      description: 管理任务, 支持暂停、恢复、终止、超时暂停自动恢复
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      - default: paused
        description: 操作项
        enum:
        - paused
        - kill
        - pause
        - resume
        in: query
        name: action
        type: string
      - default: 1m
        description: 暂停多久, 如果没设置则需要手工恢复
        in: query
        name: duration
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 管理
      tags:
      - 任务
  /api/v1/task/{task}/step/{step}:
    get:
      consumes:
      - application/json
      description: 指定任务指定步骤的执行输出, 支持WS长连接
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      - description: 步骤名称
        in: path
        name: step
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskStepLogListRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 日志
      tags:
      - 步骤
    put:
      consumes:
      - application/json
      description: 管理指定任务的指定步骤, 支持暂停、恢复、终止、超时暂停自动恢复
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      - description: 步骤名称
        in: path
        name: step
        required: true
        type: string
      - default: paused
        description: 操作项
        enum:
        - paused
        - kill
        - pause
        - resume
        in: query
        name: action
        type: string
      - default: 1m
        description: 暂停多久, 如果没设置则需要手工恢复
        in: query
        name: duration
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 管理
      tags:
      - 步骤
...
  /api/v2/task:
    post:
      consumes:
      - application/json
      description: 创建任务
      parameters:
      - description: 任务内容
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/types.STaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskCreateRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 创建
      tags:
      - 任务
  /api/v2/task/{task}:
    get:
      consumes:
      - application/json
      description: 获取任务详情
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskDetailRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 详情
      tags:
      - 任务
  /api/v2/task/{task}/step/{step}:
    get:
      consumes:
      - application/json
      description: 获取步骤详情
      parameters:
      - description: 任务名称
        in: path
        name: task
        required: true
        type: string
      - description: 步骤名称
        in: path
        name: step
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SBase-types_STaskStepRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.SBase-any'
      summary: 详情
      tags:
      - 步骤
...
swagger: "2.0"